Fix the heap-buffer-overflow issue caused by the uint32_t overflow

+++ src/codec/utils/DecodeStream.cpp
@@ -22,7 +22,7 @@
 namespace pag {
 void DecodeStream::setPosition(uint32_t value) {
   if (!checkEndOfFile(value)) {
-    positionChanged(value - _position);
+    positionChanged(static_cast<size_t>(value) - _position);
   }
 }
 
@@ -156,9 +156,9 @@ std::string DecodeStream::readUTF8String() {
     auto textLength = strlen(text);
     if (textLength > dataView.size() - _position) {
       textLength = dataView.size() - _position;
-      positionChanged(static_cast<off_t>(textLength));
+      positionChanged(textLength);
     } else {
-      positionChanged(static_cast<off_t>(textLength + 1));
+      positionChanged(textLength + 1);
     }
     return {text, textLength};
   } else {
@@ -291,14 +291,14 @@ void DecodeStream::readPoint3DList(Point3D* points, uint32_t count, float precis
   }
 }
 
-void DecodeStream::bitPositionChanged(off_t offset) {
+void DecodeStream::bitPositionChanged(size_t offset) {
   _bitPosition += offset;
   _position = BitsToBytes(_bitPosition);
 }
 
-void DecodeStream::positionChanged(off_t offset) {
+void DecodeStream::positionChanged(size_t offset) {
   _position += offset;
-  _bitPosition = static_cast<uint64_t>(_position) * 8;
+  _bitPosition = _position * 8;
 }
 
 bool DecodeStream::checkEndOfFile(uint32_t bytesToRead) {
